Avoid std::ifstream and std::ofstream.
Unfortunately std::ifstream and std::ofstream are allowed to fail in
unobvious ways when given a duff file. In particular, gcc won't error
out when a std::ifstream is created for a directory until the first read
occurs. So we write our own stream buffer classes that do error checking
and throw useful exceptions on error, and use those instead.

Backported from master. We only deprecate FDOutputStream, not kill it
completely, to avoid ABI breaks.